+++ time/src/tcp.rs
@@ -3,6 +3,7 @@
 //! This module contains shared TCP code for both server and
 //! client.
 
+use serde::{Deserialize, Serialize};
 use tokio::{
     io::{self, BufReader, ReadHalf, WriteHalf},
     net::TcpStream,
@@ -25,4 +26,14 @@ impl From<TcpStream> for TcpHandler {
         let reader = BufReader::new(reader);
         Self { reader, writer }
     }
+}
+
+/// The TCP shared configuration between clients and servers.
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
+pub struct TcpConfig {
+    /// The TCP host name.
+    pub host: String,
+
+    /// The TCP port.
+    pub port: u16,
 }